# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CEPS Portal API — referral intake, authorizations, invoices, payments, remittances, vendors, reporting
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invites:
    post:
      operationId: createInvite
      tags: [auth]
      summary: Create a portal invite (staff only) — vendor / parent-guardian / self
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteInput"
      responses:
        "201":
          description: Invite created; URL returned for staff to share
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteResult"
        "400": { description: Invalid request or linked record not found }
        "409": { description: A user with that email already exists }
  /invites/{token}:
    get:
      operationId: getInvite
      tags: [auth]
      summary: Look up an invite by token (public, no session required)
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Invite details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteInfo"
        "404": { description: Invalid or expired invite }
  /invites/{token}/accept:
    post:
      operationId: acceptInvite
      tags: [auth]
      summary: Accept an invite, set a password, and create a session (public)
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteAcceptInput"
      responses:
        "200":
          description: Account created and session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "400": { description: Invalid request }
        "404": { description: Invalid or expired invite }
        "409": { description: A user with that email already exists }
...
components:
  schemas:
...
    SessionUser:
      type: object
      required: [id, name, email, role]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role:
          type: string
          enum: [staff, service_coordinator, parent_guardian, self, vendor]
        linkedRecordId: { type: ["string", "null"] }
        linkedRecordType: { type: ["string", "null"], enum: [client, vendor, null] }
...
    InviteInput:
      type: object
      required: [email, role, linkedRecordType, linkedRecordId]
      properties:
        email: { type: string }
        role:
          type: string
          enum: [vendor, parent_guardian, self]
        linkedRecordType:
          type: string
          enum: [client, vendor]
        linkedRecordId: { type: string }
    InviteResult:
      type: object
      required: [inviteUrl]
      properties:
        inviteUrl:
          type: string
          description: "Development only: the invite link for staff to share until email is set up"
    InviteInfo:
      type: object
      required: [email, role, linkedRecordType, recordName]
      properties:
        email: { type: string }
        role:
          type: string
          enum: [vendor, parent_guardian, self]
        linkedRecordType:
          type: string
          enum: [client, vendor]
        recordName: { type: string, description: "Display name of the linked vendor/client" }
    InviteAcceptInput:
      type: object
      required: [password]
      properties:
        name: { type: string }
        password: { type: string, minLength: 8 }
...