# lib/api-zod/src/generated/types/inviteAcceptInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CEPS Portal API — referral intake, authorizations, invoices, payments, remittances, vendors, reporting
 * OpenAPI spec version: 0.1.0
 */

export interface InviteAcceptInput {
  name?: string;
  /** @minLength 8 */
  password: string;
}